refactor: 패키지 product -> products 변경
