+++ src/main/java/kitchenpos/products/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.products.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
